Update our completing-read-multiple replacement/wrapper
Among other things we use crm to read git ranges such as "A..B" and
"C...D".  As far as crm is concerned ".." and "..." are
interchangeable separators but to git they mean different things.
We cannot use `completing-read-multiple' when reading ranges because
it returns a list of read values, stripping the separators.

`magit-completing-read-multiple' was added in [1: 388a9a254] to avoid
some (additional upcoming) duplication between our two functions that
read ranges.

Giving it that name was a mistake because its function signature
is not compatible with that of `completing-read-multiple'. In
[2: cb4d59190] `magit-completing-read-multiple*' was added to fix
that, but we did not port existing callers immediately because of
the involved risk.

Meanwhile we added more kludges to `magit-completing-read-multiple'
and `magit-completing-read-multiple*', addressing additional defects
and undesirable features of various completion frameworks.

Now the time has come to deprecate `magit-completing-read-multiple' in
favor of `magit-completing-read-multiple*'.  We cannot keep using the
former name, because of the differences in the function signature.
Eventually we will inverse of direction of the deprecation.

Previously we copied `completing-read-multiple' and modified the copy
to your needs, but we then failed to incorporate changes in the
original into our copies, so now we use a wrapper instead.

Another reason to use a wrapper is that then we don't bypass any
advice that may have been added to the original.  For example the
`vertico' completion framework advices `completing-read-multiple'
but that advise obviously did not affect our copies.

1: 388a9a254f6ee985e67926d44a2d9ebad79af897
   completion: move common crm setup to a function

2: cb4d591900187b77e54ec00cac5f25b644153b5b
   magit-completing-read-multiple*: New function

Because this commit shuffles things around quite a bit, making history
inspection harder, here is a summary of all the issues that
`magit-completing-read-multiple*' addresses along with links to
earlier descriptions:

- This commit adds the `no-split' argument, which is needed when
  reading git ranges as described above.

- Until Emacs 26 `completion-pcm--all-completions' had a bug, so
  we backport the implementation from that version to Emacs 25.

- helm-crm-default-separator

- helm-completion-in-region-default-sort-fn

- ivy-sort-matches-functions-alist